+++ app/representers/stellar_base/withdrawal_request_representer.rb
@@ -8,7 +8,6 @@ module StellarBase
     property :max_amount
     property :fee_fixed
     property :fee_percent
-    property :fee_network
     property :extra_info
 
   end