+++ app/services/stellar_base/get_callback_from.rb
@@ -1,0 +1,13 @@
+module StellarBase
+  class GetCallbackFrom
+
+    def self.call(callback)
+      if callback.respond_to?(:constantize)
+        callback = callback.constantize
+      end
+
+      callback
+    end
+
+  end
+end